fixed race class selection to not flip flop when selected
user chosen race and class options are now working and are not flipping back and forth when one or the other is chosen. race and class selections are saved to a var and is ready to be added into the generator. added user clear button that removes all user entered info (name pic selections).