added vars to be passed to cyberware code
created var in cyberware function that can be passed to the cyberware placer code. this way attribute bonuses dont get added unless the cyberware is added. some things are most likely broke in the functions code, like notes still being added regardless if the cyberware is added or not. will be fixed tomorrow