added decker check for cyberware
refactored weapons section html to all be in a single parent for future coding. added an archtype check in cyberware to see if the character is a decker, if so then a datajack is automaticaly added to the character, the correct cost for yen and essence has been added. will be working on added the required cyberware for riggers soon