fixed issue in skills section
removed the ability for non casters to gain the aura reading skill. moved aura reading into the caster only section, removed its cap of garneted 6 rank and made it a 1-6 roll.  added 2 new bools for cyberware, if archtype is decker and they have a datajack then future code will lead to deck placing and buying code, if the roller gives the character a VCR then the rigging bool will be true and future code will lead to rigging pricing and such.