weapon display complete
weapon types have been fixed and everything displays correctly, weapon will not show up if there is not enough cash, that was the "issue" I thought I was having, when in fact that is the way it was programed to be in the first place. weapon added to equipment section when bought. Need to come up with adding weapons the rolled character is specialized in or knows how to use, that way the weapon bought will make since for the skills known